pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["Hugo Silva <hugofer93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "~1.1"
python-decouple = "~3.4"
SQLAlchemy = "~1.4"
Flask-WTF = "~0.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
